+++ proyecto1/database/migrations/2018_12_07_211420_create__boleta_table.php
@@ -15,7 +15,7 @@ class CreateBoletaTable extends Migration
     {
         Schema::create('boletas', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('monto_total');
+            $table->integer('monto_total');
             $table->date('fecha_pago');
             $table->unsignedInteger('id_tipo_pago');
             $table->unsignedInteger('id_reserva');
